refactor(simulator): перенести CharacterBuilder в src/simulator/
CharacterBuilder -- инфраструктура автономного симулятора (issue #2967),
а не движка: единственный фактический потребитель -- mud-sim, тесты
используют его как хелпер. Переезд в src/simulator/ убирает класс из
circle.library и приводит код в порядок перед добавлением веб-UI
поверх симулятора.

- src/engine/entities/character_builder.{h,cpp} -> src/simulator/
- namespace entities -> simulator
- CMakeLists: убран из circle.library SOURCES/HEADERS, добавлен в
  SIMULATOR_SOURCES; tests/CMakeLists добавляет .cpp в UTILITIES
- shim tests/char.utilities.hpp обновлён под новый namespace
