+++ tests/char.utilities.hpp
@@ -1,12 +1,12 @@
 #ifndef __CHAR__UTILITIES_HPP__
 #define __CHAR__UTILITIES_HPP__
 
-// Backwards-compat shim: CharacterBuilder lives in src/engine/entities/ now,
-// so it can be used by both tests and the headless balance simulator.
-#include "engine/entities/character_builder.h"
+// Backwards-compat shim: CharacterBuilder lives in src/simulator/ -- it is a
+// simulator-side helper that tests have historically reused via this alias.
+#include "simulator/character_builder.h"
 
 namespace test_utils {
-	using CharacterBuilder = entities::CharacterBuilder;
+	using CharacterBuilder = simulator::CharacterBuilder;
 }
 
 #endif // __CHAR__UTILITIES_HPP__
